apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "faker>=20.0.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.3.21",
    "anthropic>=0.7.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.24.0",
]
loadtest = [
    "locust>=2.15.0",
]
rag = [
    "sentence-transformers>=2.2.0",
    "chromadb>=0.3.21",
    "anthropic>=0.7.0",
    "numpy>=1.24.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"